trees-detection/requirements.txt
# Core dependencies
ultralytics==8.3.0
torch>=2.0.0
torchvision>=0.15.0

# Computer vision and image processing
opencv-python-headless==4.10.0.84

# Data processing and utilities
numpy>=1.21.0
scipy>=1.7.0
scikit-learn>=0.24.0

# Configuration and serialization
PyYAML==6.0.1

# Visualization
matplotlib>=3.4.0
Pillow>=8.3.0

# Optional but recommended
tensorboard>=2.10.0
tqdm>=4.62.0
requests>=2.28.0
